+++ forui/lib/src/widgets/context_menu.dart
@@ -443,7 +443,7 @@ class _State extends State<FContextMenu> with TickerProviderStateMixin {
                   onDismiss: widget.hideRegion == .none ? null : _hide,
                 ),
               ),
-        portalBuilder: (context, _) {
+        portalBuilder: (context, _, _) {
           Widget popover = Semantics(
             label: widget.semanticsLabel,
             container: true,