+++ forui/lib/src/widgets/popover/popover.dart
@@ -479,7 +479,7 @@ class _State extends State<FPopover> with TickerProviderStateMixin {
                     onDismiss: widget.hideRegion == .none ? null : _hide,
                   ),
                 ),
-          portalBuilder: (context, _) {
+          portalBuilder: (context, _, _) {
             Widget popover = Semantics(
               label: widget.semanticsLabel,
               container: true,